Revert "complete auth"
This reverts commit 9d4a20f4f6f1a24007bcbb8fb11781d68ee735e7.
