+++ App.tsx
@@ -1,14 +1,21 @@
 import React from 'react';
 import {StyleSheet,} from 'react-native';
-import { AuthProvider } from './components/auth/AuthProvider';
-import { Router } from './Router';
+
+import {NavigationContainer} from '@react-navigation/native';
+import {createNativeStackNavigator} from "@react-navigation/native-stack";
+import HomeScreen from "./containers/HomeScreen";
+
+const Stack = createNativeStackNavigator();
 
 function App() {
     return (
-        <AuthProvider>
-            <Router/>
-        </AuthProvider>
-    )
+        <NavigationContainer>
+            <Stack.Navigator>
+                <Stack.Screen name="Home" component={HomeScreen}
+                              options={{headerLargeTitle: true, headerLargeTitleShadowVisible: false}}/>
+            </Stack.Navigator>
+        </NavigationContainer>
+    );
 }
 
 export default App;
