+++ containers/HomeScreen.tsx
@@ -1,7 +1,5 @@
 import React from 'react'
 import {Button, SafeAreaView, ScrollView, StatusBar, Text, TextInput, useColorScheme, View} from "react-native";
-import { getAuthContext } from '../components/auth/AuthProvider';
-import CustomButton from '../components/CustomButton';
 import {Section} from "../components/Section";
 
 const HomeScreen = ({navigation}: any) => {
@@ -12,15 +10,15 @@ const HomeScreen = ({navigation}: any) => {
         flex: 1,
     };
 
-    const {logout, authData} = getAuthContext();
-
     return (
         <SafeAreaView style={backgroundStyle}>
             <StatusBar barStyle={isDarkMode ? 'light-content' : 'dark-content'}/>
             <ScrollView
                 contentInsetAdjustmentBehavior="automatic"
                 style={backgroundStyle}>
-                <Section title={`Welcome ${authData.username}`}></Section>
+                <Section title={"Lets get started"}>
+                    <Text>Welcome to the Goodside Developer Interview for React Native.</Text>
+                </Section>
 
                 <Section title={"UI"}>
                     <View style={{ flexDirection: "column", flex: 1 }}>
@@ -34,7 +32,6 @@ const HomeScreen = ({navigation}: any) => {
                         <View style={{ marginTop: 10, padding: 20, backgroundColor: "#ECECEC", flexDirection: "column", alignItems: "center", justifyContent: "center" }}>
                             <Text style={{ marginBottom: 5 }}>Input</Text>
                             <TextInput placeholder={"Some input"} style={{ backgroundColor: "white", padding: 10 }}/>
-                            <CustomButton text={"Log Out"} onPress={logout}/>
                         </View>
                     </View>
                 </Section>
