+++ package.json
@@ -10,12 +10,10 @@
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
   },
   "dependencies": {
-    "@react-native-async-storage/async-storage": "^1.16.1",
     "@react-navigation/native": "^6.0.7",
     "@react-navigation/native-stack": "^6.3.0",
     "react": "17.0.2",
     "react-native": "0.67.1",
-    "react-native-bouncy-checkbox": "^2.1.10",
     "react-native-safe-area-context": "^3.3.2",
     "react-native-screens": "^3.10.2"
   },
